+++ src/components/Posts.tsx
@@ -1,20 +1,34 @@
+import { useCallback, useState } from "react";
+
 import Link from "next/link";
 import Image from "next/image";
 
-import styles from "./postsList.module.scss";
+import styles from "./posts.module.scss";
+import PageNation from "./PageNation";
 
-interface Iposts {
-  posts: [];
-  offset: number;
-  PER_PAGE_COUNT: number;
-}
+export default function Posts({ posts = [] }) {
+  const [currentPage, setCurrentPage] = useState(1);
 
-export default function Posts({ posts = [], offset, PER_PAGE_COUNT }: Iposts) {
-  // pagePosts === 포스트 리스트들
-  // pageTags === 각 카테고리 개수
-  // pageTotaltags === 총 카테고리 개수
+  //pagenation => class로 바꿔보기
+  const PER_PAGE_COUNT = 6;
+  const offset = (currentPage - 1) * PER_PAGE_COUNT;
+  const totalPage = Math.ceil(posts.length / PER_PAGE_COUNT);
 
-  console.log(posts);
+  const handlePage = useCallback(
+    (move: string | undefined) => {
+      if (move === "prev") {
+        return setCurrentPage(currentPage - 1);
+      }
+
+      if (move === "next") {
+        return setCurrentPage(currentPage + 1);
+      }
+
+      return setCurrentPage(Number(move));
+    },
+    [currentPage]
+  );
+
   if (!posts.length) {
     return (
       <div style={{ backgroundColor: "black", color: "white" }}>
@@ -24,41 +38,51 @@ export default function Posts({ posts = [], offset, PER_PAGE_COUNT }: Iposts) {
   }
 
   return (
-    <ul className={styles.postList}>
-      {posts
-        .slice(offset, offset + PER_PAGE_COUNT)
-        .map(({ id, name, tags, description, createdTime }) => (
-          <li key={id}>
-            <Link href={`/posts/${id}`}>
-              <>
-                <div className={styles.imagePosition}>
-                  <Image
-                    src="http://t1.daumcdn.net/friends/prod/editor/dc8b3d02-a15a-4afa-a88b-989cf2a50476.jpg"
-                    alt="공사중"
-                    width={200}
-                    height={200}
-                  />
-                </div>
+    <>
+      <ul className={styles.postList}>
+        {posts
+          .slice(offset, offset + PER_PAGE_COUNT)
+          .map(({ id, name, tags, description, createdTime }: any) => (
+            <li key={id}>
+              <Link href={`/posts/${id}`}>
+                <>
+                  <div className={styles.imagePosition}>
+                    <Image
+                      src="http://t1.daumcdn.net/friends/prod/editor/dc8b3d02-a15a-4afa-a88b-989cf2a50476.jpg"
+                      alt="공사중"
+                      width={200}
+                      height={200}
+                    />
+                  </div>
 
-                <h2>{name}</h2>
-                <p className={styles.postMeta}>
-                  <span> 홍원배 </span>
-                  <span> | </span> <span>{createdTime}</span>
-                  <span>
-                    {tags.map(({ name }, i) => (
-                      <span key={`${name}-${i}`}>{name}</span>
-                    ))}
-                  </span>
-                </p>
-                <p className={styles.postContents}>
-                  {description.length > 85
-                    ? `${description.slice(0, 85)}...`
-                    : description}
-                </p>
-              </>
-            </Link>
-          </li>
-        ))}
-    </ul>
+                  <h2>{name}</h2>
+                  <p className={styles.postMeta}>
+                    <span> 홍원배 </span>
+                    <span> | </span> <span>{createdTime}</span>
+                    <span>
+                      {tags.map(({ name }, i: number) => (
+                        <span key={`${name}-${i}`}>{name}</span>
+                      ))}
+                    </span>
+                  </p>
+                  <p className={styles.postContents}>
+                    {description.length > 85
+                      ? `${description.slice(0, 85)}...`
+                      : description}
+                  </p>
+                </>
+              </Link>
+            </li>
+          ))}
+      </ul>
+
+      <div className={styles.pageNationContainer}>
+        <PageNation
+          currentPage={currentPage}
+          totalPage={totalPage}
+          handlePage={handlePage}
+        />
+      </div>
+    </>
   );
 }
