+++ src/components/posts.module.scss
@@ -1,0 +1,61 @@
+@use "src/styles/colors";
+
+.postList {
+  display: grid;
+  grid-template-rows: repeat(2, 1fr);
+  grid-template-columns: repeat(3, minmax(100px, 1fr));
+  row-gap: 1.2rem;
+  column-gap: 1.2rem;
+
+  li {
+    // 반응형에서 값을 바꾸도록 하자
+    height: 400px;
+    background-color: colors.$WHITE;
+    border: none;
+    border-radius: 15px;
+    transition: all 0.2s !important;
+
+    &:hover {
+      box-shadow: 0 5px 10px 3px rgba(colors.$WHITE, 0.8);
+      transform: translateY(-15px);
+    }
+
+    .imagePosition {
+      display: flex;
+      justify-content: center;
+      margin: 10px;
+    }
+
+    h2 {
+      font-size: 1.2em;
+      font-weight: bold;
+      color: colors.$BLACK;
+      text-align: center;
+    }
+
+    .postMeta {
+      margin-bottom: 20px;
+      color: colors.$BLACK;
+      text-align: right;
+
+      span {
+        margin-right: 5px;
+        font-size: 0.7em;
+        color: colors.$BLACK;
+      }
+    }
+
+    .postContents {
+      height: 150px;
+      padding: 0 15px;
+      overflow: hidden;
+      font-size: 0.8em;
+      color: colors.$BLACK;
+      word-break: break-all;
+    }
+  }
+}
+
+.pageNationContainer {
+  margin-top: 50px;
+}
